assert fix in kernel and user level (assert()s in vm.c all fail - don't use this code yet.)

+++ src/platform/i386/vm.c
@@ -46,8 +46,8 @@ kern_retype_initial(void)
 {
 	u8_t *k;
 
-	assert(mem_bootc_start() % RETYPE_MEM_NPAGES == 0);
-	assert(mem_bootc_end()   % RETYPE_MEM_NPAGES == 0);
+	assert((int)mem_bootc_start() % RETYPE_MEM_NPAGES == 0);
+	assert((int)mem_bootc_end()   % RETYPE_MEM_NPAGES == 0);
 	for (k = mem_bootc_start() ; k < mem_bootc_end() ; k += PAGE_SIZE * RETYPE_MEM_NPAGES) {
 		if (retypetbl_retype2user((void*)(chal_va2pa(k)))) assert(0);
 	}
